vgic: add vgic's interrupt type set fuction
When device is allocated to vm, some vm may want
to change the interrupt trigger type of the device.
zvm add support for this passthrough device to changing
trigger type of device's interrupt. By the way, when
individual interrupt is enable, write to type register in
GICv3 is wirte ignore.

Signed-off-by: Charlie Xiong <1981639884@qq.com>