Device: Support passthrough device model
Modified the description of virtual devices and added support
for a passthrough device model. Now there is no need to add
a new file to initialize the device every time when it is added.
On the contrary, simply add the 'ptdeviceN' node in overlay file,
and add its corresponding initialization macro in pt_device_*.c file.

Signed-off-by: Charlie Xiong <1981639884@qq.com>